fix little problems

+++ Texture2DDecoderNative/unitycrunch/crn_decomp.h
@@ -3142,7 +3142,7 @@ class crn_unpacker {
       return false;
 
     static_huffman_data_model dm[2];
-    for (uint32 i = 0; i < (has_etc_color_blocks ? 1 : 2); i++)
+    for (uint32 i = 0; i < (has_etc_color_blocks ? 1u : 2u); i++)
       if (!m_codec.decode_receive_static_data_model(dm[i]))
         return false;
 
@@ -3356,7 +3356,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x];
           uint8 endpoint_reference;
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             endpoint_reference = reference_group & 3;
             reference_group >>= 2;
@@ -3409,7 +3409,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x];
           uint8 endpoint_reference;
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             endpoint_reference = reference_group & 3;
             reference_group >>= 2;
@@ -3471,7 +3471,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x];
           uint8 endpoint_reference;
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             endpoint_reference = reference_group & 3;
             reference_group >>= 2;
@@ -3533,7 +3533,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x];
           uint8 endpoint_reference;
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             endpoint_reference = reference_group & 3;
             reference_group >>= 2;
@@ -3583,7 +3583,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x << 1];
           uint8 endpoint_reference, block_endpoint[4], e0[4], e1[4];
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             reference_group = m_codec.decode(m_reference_encoding_dm);
             endpoint_reference = (reference_group & 3) | (reference_group >> 2 & 12);
@@ -3650,7 +3650,7 @@ class crn_unpacker {
           block_buffer_element &buffer = m_block_buffer[x << 1];
           uint8 endpoint_reference, block_endpoint[4], e0[4], e1[4];
           if (y & 1) {
-            endpoint_reference = buffer.endpoint_reference;
+            endpoint_reference = (uint8)buffer.endpoint_reference;
           } else {
             reference_group = m_codec.decode(m_reference_encoding_dm);
             endpoint_reference = (reference_group & 3) | (reference_group >> 2 & 12);
